usr/projects/personal_knowledge_base/requirements.txt
# Core dependencies
numpy>=1.21.0
sqlite3

# Embeddings and ML
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# NLP for entity extraction
spacy>=3.5.0

# Web fetching
requests>=2.28.0
beautifulsoup4>=4.12.0
trafilatura>=1.6.0
aiohttp>=3.8.0
playwright>=1.40.0

# YouTube
youtube-transcript-api>=0.6.0
yt-dlp>=2023.12.0

# Twitter/X
gallery-dl>=1.25.0

# Telegram bot
python-telegram-bot>=20.0

# PDF processing
pypdf>=4.0.0
pdfplumber>=0.10.0

# Optional: for better entity extraction
# torch>=2.0.0
